Adding ghc-mod as the haskell syntax checker
Previously ghc was being used, but this was not good because
ghc would work only on individual files. As soon you included
an external module also developed on the current project, ghc would
barf at you saying that the module couldn't be found. ghc-mod
doesn't check dependencies, just syntax, also it has the lint
utility that is pretty handy.