Updating the haskell syntax checker
The previous implementation was crashing the make vim utility, after
saving it was required to use :redraw! in order to keep using the editor.

After a lot of investigation, I realized that the use of && in commands is not
of the like to the :make vim utility.

The use of the && command was updated with "{ commad1; command2 }" approach,
this way we got the result we wanted without any obnoxious side effect.