fix a small bug with the sh syntax checker and refactor
Previously, if we edited a new bash script then we would have to wipeout
the buffer and recreate it to get syntastic to recognise it as a bash
script. This is because it parsed out a magic number and cached that -
and it trivially fails to find a magic number when you create a new
file.

So recheck for a magic number if it is currently empty.

Also, do a small refactor.