less checker: update the error format for the latest version
I installed the latest version of lessc and they have modified their
output slightly. Update the errorformat for this.

Also, they new output is colorized by default. So I have added an option
to specify args to the checker that currently defaults to "--no-color".
This means that if people want to use an older version of lessc, they
will need to reset this option to an empty string.