python/flake8: assume all outputs are errors - not warnings
The output of flake8 is ambiguous as to whether results should be
interpreted as warnings or errors. I have changed it to assume errors to
match the python/pyflakes checker.

In future we may want to change this or add items to the errorformat for
some specific warnings if they are annoying.

This fixes #203.