html: split the checkers into separate files
Previously the 2 checkers were in the same file - this is more "tidy"
(lol??).

Currently the logic to determine which checker is loaded is custom to
the checker. At some point the SyntasticLoadChecker() function should be
upgraded to be able to handle cases (like this one) where the name of
the checker doesnt imply the name(s) of the executables that have to be
checked for.