Enable user-specified Ruby interpreters.
On systems where Ruby 1.8.7 is the default Ruby and users use
alternative suffixes (e.g., ruby19) or even rbenv or rvm Rubies, modern
Ruby syntax will be highlighted as an error.

The default behaviour in Syntastic is to run 'ruby' to check the syntax
of the script in question. This patch allows the user to specify a
different binary which may even be a full path. This should work on all
platforms.

  let g:syntastic_ruby_exec = 'ruby19'
  let g:syntastic_ruby_exec = '~/.rbenv/versions/1.9.2-p318/bin/ruby'