allow flake8 to work correctly with windows paths
When using syntastic with flake8 on a system with a messed up $PATH, i.e. some paths are missing (I guess), then flake8 would complain about not being able to import __main__.

This simple fix seems to fix that, as it skips the part where the *flake8* binary is executed or something. Besides that, as far as I have tested it doesn't seem to break backwards compatibility. So, if anything, there's no loss ;)