Add support for JSHint's secondary error format
JSHint has a secondary error format that includes a number related to the error message.

E.g. The normal error format is something like: 

.\Foo.js: line 2, col 5, Missing semicolon. (W033)

But the secondary error format includes a number in parentheses:

.\Foo.js: line 3, col 4, Blocks are nested too deeply. (2) (W073)"

I've changed the errorformat to read the message until it hits a space, instead of a period, so that these errors are included.