Fix errorformat for RuboCop checker
The errorformat for the RuboCop checker was missing the capture group
for the column number of the error/warning, which resulted in the
location list attempting to jump to 'filename.rb:80', which would
attempt to open a new file.

The solution was to add the "%c" to capture the column number of the
error.