Rework of the perl checker.
Helper script efm_perl.pl is no longer needed, error formatting is now
done in a preprocess function.

g:syntastic_perl_interpreter is now a list.  If set to a string,
the value is still used for backward compatibility, but a
deprecation warning is issued.  There is also a buffer-scoped
b:syntastic_perl_interpreter.

Warnings are are now checked only if no errors are found.