Prioritize `jsxhint` over `jshint`
`jsxhint` depends on having `jshint` installed, so without this
`jsxhint` will never get selected by default. If you have `jsxhint`
installed it's reasonable to assume you want to use it over `jshint`
because it will work fine on normal `.js` file as well as `.jsx` file.

Ideally `jsxhint` would only be available if the file is an actual `jsx`
file (i.e., has the `.jsx` extension or the `@jsx` pragma at the start
of the file).