Use --parse-only by default with rustc.
This reverts a previous change to use --no-trans by default when
calling the Rust compiler. --no-trans is a more useful mode, but
doesn't quite fit with Syntastic's model. Instead, --parse-only
will check that the syntax of a Rust file is valid and it will
always work. --parse-only is a more useful *default* mode.