R lint and svtools: fix CWD for Windows.
On Windows R is run with CWD set to the directory used at install time,
rather than the current directory.  The official workaround for this is
to add links to R from the home directories of each of your projects.
We can't do that in Vim, but we can call `setwd()` to Vim's idea of
current directory.  This mimics the behaviour of R on UNIX.