Add basic support for handling usage of the env binary in shebangs
Make the parseShebang function check for the usage of the env binary and
if found include the immediately proceeding argument as part of the exe
variable we return. This ensures we use the correct shell when parsing
files which use the env binary for improved portability. We don't check
for arguments to env at present, but they're rarely used, and this seems
to still be an improvement over the current situation.