Security: disable the elixir checker by default.
This executes the code your files. This is probably fine if you wrote
the files yourself, but it can be a problem if you're trying to check
third party files. If you are 100% willing to let Vim run the code in
your files, set g:syntastic_enable_elixir_checker to 1 in your vimrc.

References:

https://groups.google.com/d/msg/elixir-lang-talk/B29noPHvQ-8/9JvSGPop7n0J