Allow to run closurecompiler through shell script
Homebrew (and likely other package managers) installs a shell script in
/usr/local/bin/closure-compiler to run the installed version of the
closure-compiler. This is pretty convenient for upgrading.

This patch adds `g:syntastic_javascript_closurecompiler_executable` to
specify an executable closurecompiler rather than a path to a jar. If
this is left unspecified the closurecompiler is run with `java -jar`.