Security: disable the swift/xcrun checker by default.
The checker executes the code in your files. This is probably fine if
you wrote the files yourself, but it can be a problem if you're trying
to check third party files. If you are 100% willing to let Vim run the
code in your files, set g:syntastic_enable_swift_xcrun_checker to 1 in
your vimrc.

Reference:

https://github.com/scrooloose/syntastic/issues/1326