Add tentative ponyc checker
This patch adds support for the new [ponyc](https://github.com/CausalityLtd/ponyc) programming language.

There are some issues which can be better handled but the current result
seems to work.

I have made comments about this issue but the use of the flag `-p /usr/lib/pony`
is Arch Linux specific at the moment which will hopefully be fixed soon
so that `ponyc` will look in this location by default.