Go checker: no longer attempt to redirect output of "go build".
Running "go build -o /dev/null" in a main package that lives in GOPATH
produces an error message:

build output "/dev/null" already exists and is not an object file

This issue won't be addressed on the Go side:

http://code.google.com/p/go/issues/detail?id=4851
https://github.com/golang/go/issues/4851

Consequently, syntastic leaves this to the user to resolve.  Set
g:syntastic_go_go_build_args to something including "-o /some/dir"
to redirect the output of "go build".