Coverity desktop analysis checker
Parsing the Coverity cov-run-desktop commands on all files which were
changed according to scm from last time. It would be better, to remove
the scm flag and just pass the file to buffer, but still need to figure
out how.
Syntax marks CID's (defects) as errors, how it came to them as
information and few special misra violations as warnings.