Checker rst/rst2pseudoxml: errors without a line #
Stop a type of error failing silently; for example if a document
includes:

    `Link text`__

but misses the corresponding:

.. __: http://www.example.org

then without this patch, syntastic shows no errors. rst2pseudoxml
produces output like:

    file.txt:2: (ERROR/3) Anonymous hyperlink mismatch: 1 references but
    0 targets.
    See "backrefs" attribute for IDs.

This patch adds a pattern to 'errorformat' to catch this type of error
message that doesn't include a line number.