Security: disable the racket checker by default.
This checker executes the code in the files it checks. This is probably fine
if you wrote the files yourself, but it can be a problem if you're trying to
check third party files. If you are 100% willing to let Vim run the code in
your files, set g:enable_racket_racket_checker to 1 in your vimrc to enable
this checker: vim let g:enable_racket_racket_checker = 1 There is also a
buffer-local version of this variable, that takes precedence over it in the
buffers where it is defined.

Reference: https://github.com/scrooloose/syntastic/issues/1773