Enabling Racket checker do work without running the code
The change I propose uses `raco make` instead of just running the code. This might be a controversial change for some people because in case of smaller scripts running the code takes no time and can actually uncover more bugs. However that's not the case for every program and using raco is more universal -- will work for all files and does not carry the risk of blocking Vim by "checking" a program with long running time.
I also changed the error format to include all lines in Racket's error output. As it is the messages shown by Syntastic often arent informative enough and one must run the code again in DrRacket to get to the core of the problem.