Call uname without shell.
I don’t see a need to have uname called within a shell, calling the
program directly using the list syntax speeds up the execution of his
line by a potentially huge amount - which is good as it is run during
startup.

With shell set to `/bin/bash --login` it can take a relatively long time
to load up a login shell. Below is the difference in time for this
line during startup before and after the change, this is nearly half
a second.

count  total (s)   self (s)

1   0.479970   0.000027         let g:_SYNTASTIC_UNAME = split(syntastic#util#system('uname'), "\n")[0]

1   0.002503   0.000025         let g:_SYNTASTIC_UNAME = split(syntastic#util#system(['uname']), "\n")[0]