[clang-tidy] [clang-check] Correctly pass args
Passing the arguments like it was done is not documented and seems to
even override any options, this means that clang is in fact only invoked
with said options, which not only suppresses warnings but also ignores
input from compile_commands.json etc. including include directories.

The correct way to pass arguments to clang is to use -extra-arg=<arg>.