Checker html/w3: Add support for SVG and XHTML
Apply the same changes that were applied to html/validator in
vim-syntastic/syntastic#2241 to w3/html so that it can be used to
validate SVG and XHTML as well.

Add support for `g:syntastic_{type}_w3_doctype` to control the doctype
POST parameter sent to the validator.  This is necessary for SVG without
a DTD (which would otherwise be validated as generic XML) and useful for
all the types to control validation.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>