Adds explicit buffer number to SyntasticStatuslineFlag.
When using splits in vim, it may happen that the statusline of
an inactive window gets redrawn. In that case, the 'current()'
buffer is no longer right. This fix finds the proper buffer
number to use, by using the g:statusline_winid when it exists.
In other cases, fallback to an empty current() call.